refactor: rename backends module to bootstrappers
Rename the module directory and update references in lib/config/tests.
Refresh AGENTS.md wording to match the new bootstrappers naming.
