+++ AGENTS.md
@@ -62,7 +62,7 @@ The codebase follows a clean separation of concerns with a pluggable backend arc
   - `init_logging()`: Initializes tracing/logging with configurable levels
   - `run_apply()`: Executes the two-phase bootstrap + provisioning workflow
   - `run_validate()`: Validates profile configuration
-  - Re-exports all public modules (backends, cli, config, executor, provisioners)
+  - Re-exports all public modules (bootstrappers, cli, config, executor, provisioners)
 
 - **cli.rs**: Command-line interface definitions using clap. Defines the `Cli`, `Commands` enum (Apply, Validate, Completions), and their associated argument structs.
 
@@ -75,7 +75,7 @@ The codebase follows a clean separation of concerns with a pluggable backend arc
   - `load_profile()`: Loads and validates YAML configuration files with path resolution
   - `Profile::validate()`: Validates profile semantics including provisioner compatibility with backend output
 
-- **backends/**: Pluggable bootstrap backend implementations:
+- **bootstrappers/**: Pluggable bootstrap backend implementations:
   - **mod.rs**: Defines the `BootstrapBackend` trait and `RootfsOutput` enum
     - `BootstrapBackend` trait: `command_name()`, `build_args()`, `rootfs_output()` methods
     - `RootfsOutput`: Classifies backend output as `Directory` (provisioner-compatible) or `NonDirectory`
@@ -256,13 +256,13 @@ The provisioners system runs post-bootstrap configuration steps inside the boots
 
 ### Hook Types (mmdebstrap only)
 
-mmdebstrap supports multiple hook phases (defined in `backends/mmdebstrap.rs`):
+mmdebstrap supports multiple hook phases (defined in `bootstrappers/mmdebstrap.rs`):
 - `setup_hook`: Runs before package extraction
 - `extract_hook`: Runs after package extraction
 - `essential_hook`: Runs after essential packages are installed
 - `customize_hook`: Runs before final image creation
 
-Note: Hooks are backend-specific. If you need post-bootstrap steps that work across all backends, use provisioners instead.
+Note: Hooks are backend-specific. If you need post-bootstrap steps that work across all bootstrappers, use provisioners instead.
 
 ### Argument Building
 
@@ -270,9 +270,9 @@ Note: Hooks are backend-specific. If you need post-bootstrap steps that work acr
 Each backend originally had helper functions `add_flag()` and `add_flags()` to conditionally add arguments only when values are non-empty.
 
 **Modern Pattern (shared utility):**
-The `CommandArgsBuilder` in `backends/args.rs` provides a consistent, fluent API:
+The `CommandArgsBuilder` in `bootstrappers/args.rs` provides a consistent, fluent API:
 ```rust
-use backends::{CommandArgsBuilder, FlagValueStyle};
+use bootstrappers::{CommandArgsBuilder, FlagValueStyle};
 
 let mut builder = CommandArgsBuilder::new();
 builder.push_arg("bookworm");
@@ -282,8 +282,8 @@ let args = builder.into_args();
 ```
 
 **Migration Status:**
-- New backends should use `CommandArgsBuilder`
-- Existing backends may still use legacy helpers (refactoring welcomed but not required)
+- New bootstrappers should use `CommandArgsBuilder`
+- Existing bootstrappers may still use legacy helpers (refactoring welcomed but not required)
 
 ### Trait Object Pattern
 
@@ -322,7 +322,7 @@ for provisioner_config in &profile.provisioners {
 
 Benefits:
 - Eliminates code duplication in error handling
-- Makes adding new backends/provisioners easier
+- Makes adding new bootstrappers/provisioners easier
 - Centralizes polymorphic interaction logic in one place
 
 ### Logging
@@ -335,7 +335,7 @@ Logging is initialized once in `lib::init_logging()` and used throughout the cod
 
 ### Adding New Bootstrap Backend
 
-1. Create new backend module in `src/backends/your_tool.rs`
+1. Create new backend module in `src/bootstrappers/your_tool.rs`
 2. Define backend-specific config struct (e.g., `YourToolConfig`) with serde derives
 3. Implement the `BootstrapBackend` trait:
    ```rust
@@ -373,7 +373,7 @@ Logging is initialized once in `lib::init_logging()` and used throughout the cod
        }
    }
    ```
-6. Export the new module in `src/backends/mod.rs`:
+6. Export the new module in `src/bootstrappers/mod.rs`:
    ```rust
    pub mod your_tool;
    ```
@@ -433,7 +433,7 @@ Logging is initialized once in `lib::init_logging()` and used throughout the cod
 
 ### Adding Options to Existing Backend
 
-1. Add field to backend config struct (e.g., `MmdebstrapConfig` in `backends/mmdebstrap.rs`) with `#[serde(default)]`
+1. Add field to backend config struct (e.g., `MmdebstrapConfig` in `bootstrappers/mmdebstrap.rs`) with `#[serde(default)]`
 2. Add argument construction logic in backend's `build_args()` method:
    - Use `CommandArgsBuilder` for new code
    - Or use legacy `add_flag`/`add_flags` helpers if modifying existing code
@@ -459,7 +459,7 @@ The project has comprehensive test coverage across multiple dimensions:
 - Ensure proper handling of optional fields
 
 **Configuration Tests** (`tests/config_test.rs`):
-- Test YAML deserialization for all backends and provisioners
+- Test YAML deserialization for all bootstrappers and provisioners
 - Verify validation logic catches invalid configurations
 - Test path resolution for relative paths
 
@@ -489,7 +489,7 @@ The project has comprehensive test coverage across multiple dimensions:
 
 **Best Practices:**
 - Use helper functions from `tests/helpers/mod.rs` for consistent fixtures
-- Add new helper functions for additional backends/provisioners
+- Add new helper functions for additional bootstrappers/provisioners
 - Use the `yaml!` macro for cleaner test YAML literals
 - Use `CwdGuard` for tests that modify current directory
 - Use `CWD_TEST_LOCK` mutex for parallel test safety when changing directories
@@ -525,7 +525,7 @@ When implementing provisioners or backend hooks:
 
 - Keep backend-specific logic in backend modules
 - Keep provisioner-specific logic in provisioner modules
-- Use traits for polymorphism across backends/provisioners
+- Use traits for polymorphism across bootstrappers/provisioners
 - Prefer composition over inheritance
 - Use the builder pattern for complex construction (e.g., `CommandArgsBuilder`)
