+++ src/lib.rs
@@ -1,4 +1,4 @@
-pub mod backends;
+pub mod bootstrappers;
 pub mod cli;
 pub mod config;
 pub mod executor;
@@ -64,7 +64,8 @@ pub fn run_apply(opts: &cli::ApplyArgs, executor: &dyn CommandExecutor) -> Resul
         info!("starting provisioning phase with {} provisioner(s)", profile.provisioners.len());
 
         // Get rootfs directory (validation ensures it's a directory if provisioners exist)
-        let backends::RootfsOutput::Directory(rootfs) = backend.rootfs_output(&profile.dir)? else {
+        let bootstrappers::RootfsOutput::Directory(rootfs) = backend.rootfs_output(&profile.dir)?
+        else {
             unreachable!("validation should have caught provisioners with non-directory output")
         };
